Implements Refactored Navbar and SideBar
