+++ capstone-frontend/src/App.css
@@ -1,48 +1,12 @@
-/* .App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-} */
 
 #main-body {
   margin-top: 56px;
 }
 
 #right-sidebar {
-  margin-top: 56px;
+  margin-top: 40px;
+  margin-right: -8px;
   min-height: 100vh;   
-  background-color: #ccc;
+  background-color: whitesmoke;
+  border: 1px solid #ccc
 }
