+++ capstone-frontend/src/components/LeftSideBar.jsx
@@ -1,122 +1,453 @@
-import React, { Component } from 'react';
-import { Link, withRouter } from 'react-router-dom';
-import { Accordion, Card, Button } from 'react-bootstrap';
+import React from 'react';
+import { withRouter } from 'react-router-dom'
+
+import clsx from 'clsx';
+import { makeStyles, useTheme, fade } from '@material-ui/core/styles';
+import Drawer from '@material-ui/core/Drawer';
+import AppBar from '@material-ui/core/AppBar';
+import Avatar from '@material-ui/core/Avatar'
+import Toolbar from '@material-ui/core/Toolbar';
+import InputBase from '@material-ui/core/InputBase'
+import List from '@material-ui/core/List';
+import Collapse from '@material-ui/core/Collapse';
+import Typography from '@material-ui/core/Typography';
+import Divider from '@material-ui/core/Divider';
+import IconButton from '@material-ui/core/IconButton';
+import MenuIcon from '@material-ui/icons/Menu';
+import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
+import ChevronRightIcon from '@material-ui/icons/ChevronRight';
+import ListItem from '@material-ui/core/ListItem';
+import ListItemIcon from '@material-ui/core/ListItemIcon';
+import ListItemText from '@material-ui/core/ListItemText';
+import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction';
+import LibraryBooksIcon from '@material-ui/icons/LibraryBooks';
+import MenuBookIcon from '@material-ui/icons/MenuBook';
+import SearchIcon from '@material-ui/icons/Search';
+import HomeIcon from '@material-ui/icons/Home';
+import CollectionsIcon from '@material-ui/icons/Collections'
+import DeleteIcon from '@material-ui/icons/Delete'
+import PeopleIcon from '@material-ui/icons/People'
+import ExpandLess from '@material-ui/icons/ExpandLess';
+import ExpandMore from '@material-ui/icons/ExpandMore';
+
+
+import { Route, Link } from 'react-router-dom';
+import { Row, Col } from 'react-bootstrap'
+
+import Home from './Home';
+import Logout from './Logout';
+import Browse from './Browse';
+import MyBooks from './MyBooks';
+import MyClubs from './MyClubs';
+import { BookPage } from './BookPage';
+import ListPage from './ListPage'
+import ClubPage from './ClubPage';
+import AdminClubPage from './AdminClubPage';
+import CreateClub from './CreateClub';
 import CreateList from './CreateList'
-import "../styles/LeftSideBar.css";
+import RightSideBar from './RightSideBar';
 
+import '../styles/LeftSideBar.css'
 
-export class LeftSideBar extends Component {
+const drawerWidth = 240;
 
-  deleteBookList = async (bookListId) => {
+const useStyles = makeStyles((theme) => ({
+  root: {
+    display: 'flex',
+  },
+  appBar: {
+    zIndex: theme.zIndex.drawer + 1,
+    transition: theme.transitions.create(['width', 'margin'], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen,
+    }),
+  },
+  appBarShift: {
+    marginLeft: drawerWidth,
+    width: `calc(100% - ${drawerWidth}px)`,
+    transition: theme.transitions.create(['width', 'margin'], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen,
+    }),
+  },
+  menuButton: {
+    marginRight: 36,
+  },
+  hide: {
+    display: 'none',
+  },
+  drawer: {
+    width: drawerWidth,
+    flexShrink: 0,
+    whiteSpace: 'nowrap',
+  },
+  drawerOpen: {
+    width: drawerWidth,
+    transition: theme.transitions.create('width', {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen,
+    }),
+  },
+  drawerClose: {
+    transition: theme.transitions.create('width', {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen,
+    }),
+    overflowX: 'hidden',
+    width: theme.spacing(7) + 1,
+    [theme.breakpoints.up('sm')]: {
+      width: theme.spacing(9) + 1,
+    },
+  },
+  toolbar: {
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'flex-end',
+    padding: theme.spacing(0, 1),
+    // necessary for content to be below app bar
+    ...theme.mixins.toolbar,
+  },
+  content: {
+    flexGrow: 1,
+    padding: theme.spacing(3),
+  },
+
+  nested: {
+    paddingLeft: theme.spacing(3),
+  },
+  search: {
+    position: 'relative',
+    borderRadius: theme.shape.borderRadius,
+    backgroundColor: fade(theme.palette.common.white, 0.15),
+    '&:hover': {
+      backgroundColor: fade(theme.palette.common.white, 0.25),
+    },
+    marginRight: theme.spacing(2),
+    marginLeft: 0,
+    width: '100%',
+    [theme.breakpoints.up('sm')]: {
+      marginLeft: theme.spacing(3),
+      width: 'auto',
+    },
+  },
+  searchIcon: {
+    padding: theme.spacing(0, 2),
+    height: '100%',
+    position: 'absolute',
+    pointerEvents: 'none',
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  inputRoot: {
+    color: 'inherit',
+  },
+  inputInput: {
+    padding: theme.spacing(1, 1, 1, 0),
+    // vertical padding + font size from searchIcon
+    paddingLeft: `calc(1em + ${theme.spacing(4)}px)`,
+    transition: theme.transitions.create('width'),
+    width: '100%',
+    [theme.breakpoints.up('md')]: {
+      width: '20ch',
+    },
+  },
+  grow: {
+    flexGrow: 1,
+  },
+  sectionDesktop: {
+    display: 'none',
+    [theme.breakpoints.up('md')]: {
+      display: 'flex',
+    },
+  },
+  sectionMobile: {
+    display: 'flex',
+    [theme.breakpoints.up('md')]: {
+      display: 'none',
+    },
+  },
+  small: {
+    width: theme.spacing(4),
+    height: theme.spacing(4),
+  },
+}));
+
+export const LeftSideBar = withRouter((props) => {
+
+  const classes = useStyles();
+  const theme = useTheme();
+  const [open, setOpen] = React.useState(false);
+  const [openList, setOpenList] = React.useState(true);
+  const profileObj = JSON.parse(window.localStorage.getItem("profileObj")) || {};
+
+  const handleDrawerOpen = () => {
+    setOpen(true);
+  };
+
+  const handleDrawerClose = () => {
+    setOpen(false);
+  };
+
+  const handleClick = () => {
+    setOpenList(!openList);
+  };
+
+  const handleSearchInput = (event, props) => {
+    event.preventDefault();
+    const searchValue = document.getElementById('search-input').value;
+    props.setSearchQuery(searchValue);
+    props.history.push(`/browse/${searchValue}`);
+  }
+
+  const deleteBookList = async (bookListId, props, currentUrlPath) => {
 
     // Delete BookList in Firebase
     await fetch(`/api/booklist?id=${bookListId}`, {
       method: "DELETE",
     });
 
-    this.props.updateBookLists();
+    props.updateBookLists();
 
     // These lines of code check if the current BookList URL has the same id as the one being deleted
     // If so we redirect home
-    const currentUrlPath = new URL(window.location.href).pathname;
 
     if (currentUrlPath.includes(bookListId)) {
-      this.props.history.push("/")
+      props.history.push("/")
     }
-  }
+    else {
+      props.history.push(currentUrlPath)
+    }
+  };
 
-  render() {
-    return (
-      <div id="left-sidebar-container" className="sidebar-expanded d-none d-md-block col-2 sidebar-container">
-        <ul className="list-group sticky-top sticky-offset">
+  return (
+    <div className={classes.root}>
+      <AppBar
+        style={{ backgroundColor: '#007bff' }}
+        position="fixed"
+        className={clsx(classes.appBar, {
+          [classes.appBarShift]: open,
+        })}
+      >
+        <Toolbar>
+          <IconButton
+            color="inherit"
+            aria-label="open drawer"
+            onClick={handleDrawerOpen}
+            edge="start"
+            className={clsx(classes.menuButton, {
+              [classes.hide]: open,
+            })}
+          >
+            <MenuIcon />
+          </IconButton>
 
-          <Accordion defaultActiveKey="0" style={{ marginRight: -14.3 }}>
-            <Card>
-              <Link to="/" className="bg-dark list-group-item list-group-item-action">
-                <div className="d-flex w-100 justify-content-start align-items-center">
-                  <span className="fa fa-tasks fa-fw mr-3"></span>
-                  <span id="home-link" className="menu-collapsed"> Home </span>
-                </div>
-              </Link>
-            </Card>
+          <div className={classes.sectionDesktop}>
+            <Link to="/" id='bookbook-btn'>
+              <Typography variant="h6" noWrap>
+                BookBook
+              </Typography>
+            </Link>
+          </div>
 
-            <Card>
-              <Link to="/browse" className="bg-dark list-group-item list-group-item-action">
-                <div className="d-flex w-100 justify-content-start align-items-center">
-                  <span className="fa fa-tasks fa-fw mr-3"></span>
-                  <span id="browse-link" className="menu-collapsed"> Browse </span>
-                </div>
-              </Link>
-            </Card>
+          <div className={classes.search}>
+            <div className={classes.searchIcon}>
+              <SearchIcon />
+            </div>
+            <form onSubmit={(event) => handleSearchInput(event, props)} className="remove-form-style" autoComplete="off">
+              <InputBase
+                autoComplete="off"
+                id="search-input"
+                placeholder="Search…"
+                classes={{
+                  root: classes.inputRoot,
+                  input: classes.inputInput,
+                }}
+                inputProps={{ 'aria-label': 'search' }}
+              />
+            </form>
+          </div>
+          <div className={classes.grow} />
 
-            <Card>
-              <Accordion.Toggle as={Card.Header} eventKey="2">
-                <Link to="/myreads" className="bg-dark list-group-item list-group-item-action">
-                  <div className="d-flex w-100 justify-content-start align-items-center">
-                    <span className="fa fa-tasks fa-fw mr-3"></span>
-                    <span id="mybooks-link" className="menu-collapsed"> My Books </span>
-                  </div>
-                </Link>
-              </Accordion.Toggle>
-              <Accordion.Collapse eventKey="2">
-                <Card.Body>My Books Dropdown</Card.Body>
-              </Accordion.Collapse>
-            </Card>
+          <div className={classes.sectionDesktop}>
+            <Logout toggleSignIn={props.toggleSignIn} />
+          </div>
+        </Toolbar>
+      </AppBar>
+      <Drawer
+        variant="permanent"
+        className={clsx(classes.drawer, {
+          [classes.drawerOpen]: open,
+          [classes.drawerClose]: !open,
+        })}
+        classes={{
+          paper: clsx({
+            [classes.drawerOpen]: open,
+            [classes.drawerClose]: !open,
+          }),
+        }}
+      >
+        <div className={classes.toolbar}>
 
-            <Card>
-              <Accordion.Toggle as={Card.Header} eventKey="3">
-                <div className="bg-dark list-group-item list-group-item-action">
-                  <div className="d-flex w-100 justify-content-start align-items-center">
-                    <span className="fa fa-list fa-fw mr-3"></span>
-                    <span> My Library </span>
-                  </div>
-                </div>
-              </Accordion.Toggle>
+          <div className={classes.sectionDesktop}>
+            <Avatar
+              alt={profileObj.name}
+              src={profileObj.imageUrl} />
 
-              <Accordion.Collapse eventKey="3">
-                <Card.Body id="mylists-link">
-                  {
-                    this.props.bookLists.map(bookList =>
-                      <div key={bookList.id} className="bg-light p-0 list-group-item list-group-item-action">
-                        <div className="d-flex w-100 justify-content-start align-items-center">
-                          <Link to={`/listpage/${bookList.id}`} className="bg-light border-0 list-group-item list-group-item-action">
-                            <span> {bookList.name}</span>
-                          </Link>
-                          <Button className="border-0 bg-transparent" onClick={() => this.deleteBookList(bookList.id)}>
-                            <svg width="1.5em" height="1.5em" viewBox="0 0 16 16" className="bi bi-trash-fill" fill="black" xmlns="http://www.w3.org/2000/svg">
-                              <path fillRule="evenodd" d="M2.5 1a1 1 0 0 0-1 1v1a1 1 0 0 0 1 1H3v9a2 2 0 0 0 2 2h6a2 2 0 0 0 2-2V4h.5a1 1 0 0 0 1-1V2a1 1 0 0 0-1-1H10a1 1 0 0 0-1-1H7a1 1 0 0 0-1 1H2.5zm3 4a.5.5 0 0 1 .5.5v7a.5.5 0 0 1-1 0v-7a.5.5 0 0 1 .5-.5zM8 5a.5.5 0 0 1 .5.5v7a.5.5 0 0 1-1 0v-7A.5.5 0 0 1 8 5zm3 .5a.5.5 0 0 0-1 0v7a.5.5 0 0 0 1 0v-7z" />
-                            </svg>
-                          </Button>
-                        </div>
-                      </div>
-                    )
-                  }
+            <Typography
+              noWrap
+              gutterBottom
+              className="avatar-text-style"
+              align="center"
+              color="textPrimary"
+              variant="h6" >
+              {profileObj.name}
+            </Typography>
+          </div>
 
-                  <CreateList updateBookLists={this.props.updateBookLists} btnStyle="bg-light list-group-item list-group-item-action" textStyle={"d-flex w-100 justify-content-start align-items-center"}/>
+          <div className={classes.sectionMobile.concat(' margin-auto')}>
+            <Avatar
+              className={classes.small}
+              alt={profileObj.name}
+              src={profileObj.imageUrl} />
 
-                </Card.Body>
-              </Accordion.Collapse>
-            </Card>
+            <Typography
+              noWrap
+              gutterBottom
+              className="avatar-text-style"
+              align="center"
+              color="textPrimary"
+              variant="h6" >
+              {profileObj.name}
+            </Typography>
+          </div>
 
-            <Card>
-              <Accordion.Toggle as={Card.Header} eventKey="4">
-                <Link to="/myclubs" className="bg-dark list-group-item list-group-item-action">
-                  <div className="d-flex w-100 justify-content-start align-items-center">
-                    <span className="fa fa-tasks fa-fw mr-3"></span>
-                    <span id="myclubs-link" className="menu-collapsed"> My Clubs </span>
-                  </div>
-                </Link>
-              </Accordion.Toggle>
-              <Accordion.Collapse eventKey="4">
-                <Card.Body>My Clubs Dropdown</Card.Body>
-              </Accordion.Collapse>
-            </Card>
-          </Accordion>
+          <IconButton onClick={handleDrawerClose}>
+            {theme.direction === 'rtl' ? <ChevronRightIcon /> : <ChevronLeftIcon />}
+          </IconButton>
 
-        </ul>
-      </div >
-    );
-  }
-}
+        </div>
 
-export default withRouter(LeftSideBar)
+        <Divider />
+        <List>
+          <Link to="/" className="remove-link-style">
+            <ListItem button>
+
+              <ListItemIcon>
+                <HomeIcon />
+              </ListItemIcon>
+              <ListItemText primary="Home" />
+
+            </ListItem>
+          </Link>
+
+          <Link to="/browse" className="remove-link-style">
+            <ListItem button>
+
+              <ListItemIcon>
+                <SearchIcon />
+              </ListItemIcon>
+              <ListItemText primary="Search" />
+
+            </ListItem>
+          </Link>
+
+          <Link to="/myreads" className="remove-link-style">
+            <ListItem button >
+
+              <ListItemIcon>
+                <MenuBookIcon />
+              </ListItemIcon>
+              <ListItemText primary="My Books" />
+
+            </ListItem>
+          </Link>
+
+          <ListItem button onClick={handleClick}>
+
+            <ListItemIcon>
+              <LibraryBooksIcon />
+            </ListItemIcon>
+            <ListItemText primary="My BookLists" />
+
+            {openList ? <ExpandLess /> : <ExpandMore />}
+          </ListItem>
+
+          <Collapse in={openList} timeout="auto" unmountOnExit>
+            <List component="div" disablePadding>
+              {
+                props.bookLists.map(bookList =>
+
+                  <Link to={`/listpage/${bookList.id}`} key={bookList.id} className="remove-link-style">
+                    <ListItem button className={classes.nested} >
+
+                      <ListItemIcon>
+                        <CollectionsIcon />
+                      </ListItemIcon>
+
+                      <ListItemText primary={bookList.name} />
+
+                      {open &&
+                        <ListItemSecondaryAction>
+                          <IconButton onClick={() => deleteBookList(bookList.id, props, window.location.pathname)}>
+                            <DeleteIcon />
+                          </IconButton>
+                        </ListItemSecondaryAction>
+                      }
+                    </ListItem>
+                  </Link>
+                )
+              }
+              <CreateList updateBookLists={props.updateBookLists} sideBar={true} />
+            </List>
+          </Collapse>
+
+          <Link to="/myclubs" className="remove-link-style">
+            <ListItem button >
+
+              <ListItemIcon>
+                <PeopleIcon />
+              </ListItemIcon>
+              <ListItemText primary="My Clubs" />
+
+            </ListItem>
+          </Link>
+
+          {open &&
+            <>
+              <Divider />
+              <ListItem button >
+                <Logout loginStyle={{ margin: 'auto' }} toggleSignIn={props.toggleSignIn} className="margin-auto" />
+              </ListItem>
+            </>
+          }
+
+        </List>
+      </Drawer>
+      <main className={classes.content}>
+        <Row>
+          <Col id="main-body">
+            <Route exact path='/' component={Home} />
+            <Route path='/browse/:query' render={(pageProps) => (
+              <Browse bookLists={props.bookLists} updateBookLists={props.updateBookLists} searchQuery={pageProps.match.params.query} />
+            )} />
+            <Route path='/mybooks' component={MyBooks} />
+            <Route path='/listpage/:id' component={ListPage} />
+            <Route path='/myclubs' component={MyClubs} />
+            <Route path='/bookpage/:id' render={(pageProps) => (
+              <BookPage bookId={pageProps.match.params.id} bookLists={props.bookLists} updateBookLists={props.updateBookLists} />
+            )} />
+            <Route path='/clubpage/:id' render={(pageProps) => (
+              <ClubPage id={pageProps.match.params.id} bookLists={props.bookLists} updateBookLists={props.updateBookLists} />
+            )} />
+            <Route path='/createclub' component={CreateClub} />
+          </Col>
+          {/* TODO(#86) Add Friends to the Web App and display them in the Right Side Bar */}
+          {/* <div className={classes.sectionDesktop}>
+            <RightSideBar />
+          </div> */}
+        </Row>
+      </main>
+    </div >
+  );
+})
