--- capstone-frontend/src/components/RightSideBar.js
@@ -1,18 +1,0 @@
-import React from 'react';
-import '../App.css';
-
-const RightSideBar = () => {
-  return (
-    <div>
-    <h2>THE BODY</h2>
-    <br/>
-
-    <br/><br/><br/><br/><br/><br/><br/><br/><br/><br/>
-    </div>
-    // <div id="right-sidebar-container" className="sidebar-expanded d-none d-md-block col-2 sidebar-container">
-    //   Space for a user feed!
-    // </div>
-  );
-};
-
-export default RightSideBar;
