+++ capstone-frontend/src/styles/LeftSideBar.css
@@ -15,3 +15,17 @@
   display: inherit;
   text-decoration: none;
 }
+
+#bookbook-btn {
+  cursor: pointer;
+  color: inherit;
+  text-decoration: none;
+}
+
+.avatar-text-style {
+  font-size: 0.9rem !important;
+  font-weight: 400 !important;
+  margin-top: 0.4rem !important;
+  padding-left: 1rem;
+  padding-right: 1rem;
+}
